+++ CHANGELOG.md
@@ -6,9 +6,11 @@
     - Removing & Replacing Elements
     - Event Listeners & The Event Object
     - Mouse Events
+    - Keyboard & Input Events
 - An example with removing and replacing of elements was added.
 - An example with event listeners was added.
 - An example with mouse events was added.
+- An example with keyboard and input events was added.
 
 ## August 18, 2020
 - Modern JavaScript From The Beginning (Udemy)